Remove options.configuration DiffSuppressFunc.
With this DiffSuppressFunc in place changes to options.configuration
values will not be detected. Consider the following use case:
1. Set foo=bar and perform terraform apply.
2. When the value for foo=bar is retrieved next, it will be
encrypted as "2.0$abc123"
3. set foo=baz and perform terraform apply. Because the current,
encrypted value is "2.0$abc123" no diff will be detected.

With the current form the old, encrypted value of "2.0$abc123" will
always be different than the new, plaintext value. So terraform will
always detect a change when terraform apply is run. However, the
remote Auth0 configuration is in a consistent state.