Use set instead of list for enabled clients
The `schema.Set` type is intended for unordered lists, so using it means
we don't need to define a `DiffSuppressFunc` to prevent differences from
appearing due to the ordering of the enabled clients.

The `DiffSuppressFunc` was causing incorrect behaviour when adding a new
enabled client to a connection so rather than fix the function using a
set seems simpler.