Return slice from `Set()`
As pointed out in PR #105, previously, if the field hasn't changed, the
`Set()` function would return a nil pointer so calling `Set().List()`
would cause an exception.

Instead, we return a slice from the function, which will be empty if
there was no change, the value didn't successfully cast to a
`*schema.Set`, or if there's no value.  We then use this value directly
to set `EnabledClients`, rather than calling the `List()` function.